refactored a lot and changed names to be more correct
