+++ Assets/Scripts/ControllerInteractions.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: fce500faaffe7ac4a9135cb03c53891a
-timeCreated: 1498068661
-licenseType: Free
+guid: 8a6b3d5534d58a1458d4e900f116c673
+timeCreated: 1520572705
+licenseType: Pro
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
