+++ Assets/Scripts/MapInteractions.cs.meta
@@ -1,9 +1,8 @@
 fileFormatVersion: 2
-guid: 8a6b3d5534d58a1458d4e900f116c673
-timeCreated: 1520572705
-licenseType: Pro
+guid: fce500faaffe7ac4a9135cb03c53891a
+timeCreated: 1498068661
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
